chore: refactor all source modules
