+++ sage/app.py
@@ -1,6 +1,6 @@
 import asyncio
 
-from sage.utils.sources import Source
+from sage.sources.sources import Source
 
 # myjira = Jira()
