+++ sage/sources/source_manager.py
@@ -19,7 +19,7 @@ from sage.constants import (
     sources_config,
 )
 from sage.utils.exceptions import SourceException
-from sage.utils.loaders import CustomConfluenceLoader, GitlabLoader, WebLoader
+from sage.sources.loaders import CustomConfluenceLoader, GitlabLoader, WebLoader
 from sage.utils.supports import CustomFAISS as FAISS
 from sage.validators.config_toml import ConfluenceModel, Files, GitlabModel, Web
