+++ sage/sources/utils.py
@@ -10,7 +10,7 @@ from langchain.schema.document import Document
 from langchain.schema.runnable import RunnableLambda
 
 from sage.constants import SENTINEL_PATH, logger
-from sage.utils.sources import Source
+from sage.sources.sources import Source
 
 
 def format_docs(docs: Sequence[Document]) -> str:
